+++ WORKSPACE
@@ -21,13 +21,18 @@ nodejs_register_toolchains(
     node_version = DEFAULT_NODE_VERSION,
 )
 
-load("//jasmine:repositories.bzl", "jasmine_repositories")
+load("@aspect_rules_js//npm:npm_import.bzl", "npm_translate_lock")
 
-jasmine_repositories(name = "jasmine")
+npm_translate_lock(
+    name = "npm",
+    npmrc = "//:.npmrc",
+    pnpm_lock = "//:pnpm-lock.yaml",
+    verify_node_modules_ignored = "//:.bazelignore",
+)
 
-load("@jasmine//:npm_repositories.bzl", jasmine_npm_repositories = "npm_repositories")
+load("@npm//:repositories.bzl", "npm_repositories")
 
-jasmine_npm_repositories()
+npm_repositories()
 
 # For running our own unit tests
 load("@bazel_skylib//:workspace.bzl", "bazel_skylib_workspace")
