+++ e2e/jasmine_test/.bazelignore
@@ -1,0 +1,1 @@
+node_modules
