+++ e2e/jasmine_test/pnpm-workspace.yaml
@@ -1,0 +1,2 @@
+packages:
+  - "."
