+++ e2e/workspace/.bazelignore
@@ -1,0 +1,1 @@
+node_modules
