+++ e2e/workspace/package.json
