+++ jasmine/BUILD.bazel
@@ -4,17 +4,6 @@ load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
 exports_files(["defs.bzl"])
 
 bzl_library(
-    name = "repositories",
-    srcs = ["repositories.bzl"],
-    visibility = ["//visibility:public"],
-    deps = [
-        "//jasmine/private:versions",
-        "@bazel_tools//tools/build_defs/repo:http.bzl",
-        "@bazel_tools//tools/build_defs/repo:utils.bzl",
-    ],
-)
-
-bzl_library(
     name = "dependencies",
     srcs = ["dependencies.bzl"],
     visibility = ["//visibility:public"],
