+++ jasmine/private/BUILD.bazel
@@ -1,19 +1,11 @@
 load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
 
 exports_files([
-    "junit_reporter.js",
-    "package.json",
-    "runner.js",
+    "runner.cjs",
+    "junit_reporter.cjs",
 ])
 
 # gazelle:exclude v*.*.*/*
-
-bzl_library(
-    name = "versions",
-    srcs = ["versions.bzl"],
-    visibility = ["//jasmine:__subpackages__"],
-    deps = [],
-)
 
 bzl_library(
     name = "jasmine_test",
@@ -22,5 +14,8 @@ bzl_library(
         "jasmine_test.bzl",
     ],
     visibility = ["//jasmine:__subpackages__"],
-    deps = ["@aspect_rules_js//js:libs"],
+    deps = [
+        "@aspect_bazel_lib//lib:copy_file",
+        "@aspect_rules_js//js:libs",
+    ],
 )
