--- jasmine/private/v4.5.0/package.json
@@ -1,6 +1,0 @@
-{
-	"dependencies": {
-		"jasmine": "4.5.0",
-		"jasmine-reporters": "2.5.2"
-	}
-}
