--- jasmine/private/versions.bzl
@@ -1,8 +1,0 @@
-"""List of mirrored versions"""
-
-# Run /scripts/mirror_release.sh to add new versions.
-# Versions should be descending order so TOOL_VERSIONS[0] is the latest version.
-TOOL_VERSIONS = [
-    "v4.5.0",
-    "v4.3.0",
-]
