--- jasmine/tests/versions_test.bzl
@@ -1,18 +1,0 @@
-"""Unit tests for starlark helpers
-See https://docs.bazel.build/versions/main/skylark/testing.html#for-testing-starlark-utilities
-"""
-
-load("@bazel_skylib//lib:unittest.bzl", "asserts", "unittest")
-load("//jasmine/private:versions.bzl", "TOOL_VERSIONS")
-
-def _smoke_test_impl(ctx):
-    env = unittest.begin(ctx)
-    asserts.equals(env, "v4.5.0", TOOL_VERSIONS[0])
-    return unittest.end(env)
-
-# The unittest library requires that we export the test cases as named test rules,
-# but their names are arbitrary and don't appear anywhere.
-_t0_test = unittest.make(_smoke_test_impl)
-
-def versions_test_suite(name):
-    unittest.suite(name, _t0_test)
