+++ package.json
@@ -1,6 +1,5 @@
 {
   "private": true,
-  "type": "module",
   "devDependencies": {
     "jasmine": "4.6.0",
     "jasmine-core": "4.6.0",
